Iterate styling of presence indicators, various refactoring
